+++ pymc3/step_methods/metropolis.py
@@ -2,7 +2,7 @@ import numpy as np
 import numpy.random as nr
 import theano
 import scipy.linalg
-import warnings
+import logging
 
 from ..distributions import draw_values
 from .arraystep import ArrayStepShared, PopulationArrayStepShared, ArrayStep, metrop_select, Competence
@@ -13,9 +13,9 @@ __all__ = ['Metropolis', 'BinaryMetropolis', 'BinaryGibbsMetropolis',
            'CategoricalGibbsMetropolis', 'NormalProposal', 'CauchyProposal',
            'LaplaceProposal', 'PoissonProposal', 'MultivariateNormalProposal']
 
+_log = logging.getLogger('pymc3')
+
 # Available proposal distributions for Metropolis
-
-
 class Proposal(object):
     def __init__(self, s):
         self.s = s
@@ -516,7 +516,7 @@ class DEMetropolis(PopulationArrayStepShared):
 
     def __init__(self, vars=None, S=None, proposal_dist=None, lamb=None, scaling=0.001,
                  tune=True, tune_interval=100, model=None, mode=None, **kwargs):
-        warnings.warn('Population based sampling methods such as DEMetropolis are experimental.' \
+        _log.warning('Population based sampling methods such as DEMetropolis are experimental.' \
             ' Use carefully and be extra critical about their results!')
 
         model = pm.modelcontext(model)